+++ internal/pkg/model/message/message.go
@@ -1,7 +1,7 @@
 package message
 
 type Message struct {
-	Source  uint16      `json:"source"`
+	Source  int      `json:"source"`
 	Type    Type        `json:"type"`
 	Payload interface{} `json:"payload"`
 }
